Reed suggested Update README.adoc
I took a look at the template because I am now editing the GUF Conceptual Model. AS you know, I am also editing the ModSpec. I made a number of suggested changes to the template description to be consistent with the ModSpec. For example, the ModSpec states that each requirement shall have a corresponding conformance test.